Added count to TransactionStateMessage.

+++ src/main/scala-2.11/com/bwsw/tstreams/agents/consumer/subscriber_v2/TransactionState.scala
@@ -10,7 +10,7 @@ import com.bwsw.tstreams.coordination.messages.state.TransactionStatus.ProducerT
 case class TransactionState(uuid: UUID,
                             partition: Int,
                             masterSessionID: Int,
-                            queueOrderID: Int,
+                            queueOrderID: Long,
                             itemCount: Int,
                             state: ProducerTransactionStatus,
                             ttl: Int)
